Factor out annotations codec
